Update 100-atoi.c long

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -9,7 +9,7 @@
 int _atoi(char *s)
 {
 	char sign;
-	int n;
+	long long int n;
 
 	sign = '+';
 	n = 0;
